splits affinity
